requirements.txt
Django==1.10.1
djangorestframework==3.4.7
pypostal==0.3
